+++ index.html
@@ -12,6 +12,6 @@
       <div class="grid"></div>
     </div>
 
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </body>
 </html>
